update
git-svn-id: https://svn.oss.deltares.nl/repos/imod/trunk@2948 8a8853d3-af7d-47ec-b359-4bd2096ba547

+++ MODFLOW/MF2005/gwf2bcf7.f
@@ -1115,7 +1115,7 @@ C     ------------------------------------------------------------------
 
 c arguments
       integer, intent(in) :: k,iminkd                                          ! DLT
-      real,intent(in) :: minkd,trpy
+      real(kind=8),intent(in) :: minkd,trpy
 !      real, dimension(ncol,nrow,nlay), intent(inout) :: cc              ! DLT
 !      real, dimension(ncol,nrow,nlay), intent(inout) :: cr              ! DLT
 C     ------------------------------------------------------------------
